07-streaming/workshop/pyproject.toml
[project]
name = "workshop"
version = "0.1.0"
description = "PyFlink Stream Processing Workshop"
requires-python = ">=3.9"
dependencies = [
    "kafka-python-ng>=2.2.3",
    "psycopg2-binary>=2.9.11",
]
